主键ID统一为Long类型

+++ src/main/java/com/yunche/loan/domain/dataObj/BizModelRelaPartnersDOKey.java
@@ -3,7 +3,7 @@ package com.yunche.loan.domain.dataObj;
 public class BizModelRelaPartnersDOKey {
     private Long bizId;
 
-    private Integer groupId;
+    private Long groupId;
 
     public Long getBizId() {
         return bizId;
@@ -13,11 +13,11 @@ public class BizModelRelaPartnersDOKey {
         this.bizId = bizId;
     }
 
-    public Integer getGroupId() {
+    public Long getGroupId() {
         return groupId;
     }
 
-    public void setGroupId(Integer groupId) {
+    public void setGroupId(Long groupId) {
         this.groupId = groupId;
     }
 }
